+++ CHANGELOG.md
@@ -9,7 +9,11 @@
   - No need to update MOP in order to add mod compatibility! (At least in most cases)
   - Mod compatibility rules can be updated if needed! MOP checks for rules update every week (you can force new check in the settings)
 - Engine renderers are now disabled, if player is in the car, and hood is attached
-- Added disclaimer for pirated game copies
+- Added disclaimer for pirated copies of the game
+
+### Bug Fixes
+
+- Boat will not disappear/teleport back to the spawn position on respawn
 
 ## 2.4.3 (19.03.2020)